Fibonacci Sequence 3 Different Ways
So this code is a mixture of code from class, code I worked on myself, and code that I had found useful on the internet. I had built the main function by myself as it was easy, worked with my class (even though I could have done it myself) to make a recursive and iterative approach to finding the number in the Nth placement of the Fibonacci sequence. To better help understand the third attempt I looked online and found a matrix approach to this sequence, using the matrix [1, 1, 1, 0]. The overall lesson supposed to be learned is that it takes longer for a program to run using a recursive approach. (Also yes, I am aware that the time is not measured in milliseconds)